Add nemo-metadata-tab (remove nemo-audio-tab)
nemo-metadata-tab adds a property page for media files showing their
metadata.
Unlike nemo-audio-tab, which becomes obselete and hence was removed, the
new metadata-tab shows metadata from audio, video, image and PDF files.
External libraries used:
 - for audio: taglib
 - for video: MediaInfoDLL3
 - for images: exifread (by @ianare, included here, BSD license)

PDF files are processed reliably via poppler's pdfinfo.
All default tags and values are being translated before showing them to
the user, all non-default tags are displayed as is, providing further
information (The latter are hidden in a Gtk Expander on load).
The extension processes and displays cover images in ID3 and FLAC-based
files and displays them in the property page. If none is found and for
other media types, the system's generic MIME type icon is shown.
The property page style was adopted to other pages' representations.

Note that nemo itself provides a property page for images' metadata.
Either that or this new implementation can and should be disabled. This
extension provides an option for that purpose in line 34.