Remove nemo-rabbitvcs
No matter how you look at it, a total rewrite is needed to enable
Git/SVN integration into nemo with Python 3. Not only the extension but
also RabbitSVN itself is in no good shape and it is simply to much work
to make it even *run somehow* with Python. So I recommend to just drop
it here.
Whether a file browser is a good place to do checkouts/commits with is
surely a discussion point itself. Nevertheless I suggest writing a new
extension sometime for users who are fond of such.