compare: Refactor nemo-compare
Nemo Compare (and its scripts) now follow the current Python design
guidelines. This enhances readability and is easier to maintain. Plus,
we have a minor performance gain.
The configuration file format was changed slightly, so the tool now
checks whether a file with the old format is present and replaces it
with default settings, so one may have to edit the configuration one
time (with the -preferences script).
Also there is a small improvement: Now it is also possible to remember
multiple files and then compare them to multiple other ones.