compare: Set new version in setup.py
As nemo-compare underwent many changes and is considered refactored for
future releases, set the package version to 3.0.0.
Also this helps distinguish this version from prior ones built for the
python2 base.

Note that the changelog file in debian/ has to be edited before packages
can be built with this new version number.