metadata-*: Fix hangs when processing (many) image stubs
As GdkPixbuf.Pixbuf.new_from_file() which we used for basic processing
on image metadata stubs turns out to be very slow, we let PIL do the
job now.
This performance issue emerged mainly in nemo-metadata-columns in
directories with many such files, but nemo-metadata-tab benefits from
this change, too.
Also, we can retrieve the image resolution easily this way, so add that
to the information.