terminal: Refactor nemo-terminal
This commit attempts to refactor nemo-terminal in order to make it
slightly easier to read and maintain.
The extensive use of copyright information was nearly left unchanged
while trying to adept to the code and distribution style of other
nemo-externsions.

NOT addressed was the issue that this extension produces many Gtk
warnings on window close.
Also a preferences tool is still missing.

Many commented lines were removed as well as the nautilus image (the
About dialog now shows a generic terminal symbol). Translation files
were introduced.
All debian received updates and a pybuild file was created.
Only the debian/changelog file has still to updated to reflect the
current version number.