pastebin: Refactor nemo-pastebin
Nemo Pastebin now follows the current Python design guidelines. This
enhances readability and is easier to maintain. Plus, we have a minor
performance gain due to simpler functions and less lookups.

Notifications are now an essential part of the extension as they provide
visual feedback for the operation progresses. Also, the new version adds
actions to the notifications, so the user can interact directly with the
backend.
Moreover we deliver now an improved version of the preferences tool and
support more options to be passed to pastebinit.
Although the code of the underlying pastebinit backend script seems in
need of an overhaul, it still suits our needs for now.
In order to keep the user updated about the current pastebin selected in
the configuration, pyinofity is used where available to watch the config
file constantly and initiate reloads without having to restart nemo.

All debian and pybuild files were updated.
Only the debian/changelog file has still to updated to reflect the
current version number.

Python version 3.3 is needed for the timeout option in subprocess. If
you need to provide backwards compatibility (down to 3.1), remove that
option and change the exception to subprocess.CalledProcessError.