python: Add import sanitizer back
Previously, the import sanitizer was removed in 141bc98 as it broke all
imports in all extension scripts due to unknown reasons. The reason for
that issue is now resolved: Python 3 changed the filter function to
being a generator. Hence, reading from the changed sys.path more than
once led to an empty list. Replacing the filter function with a list
comprehension gives us the desired behavior.