python: Fix multiarch libs
As the path where Python scripts for nemo-extensions are looked for
differs between plain Ubuntu and Linux Mint, we use a more general
approach for determining the libdir.

For PPA/Ubuntu users a postinst script now checks whether the
extensionsdir was set correctly during compilation and symlinks to the
corrected version on the current system if necessary.

m4/python.m4 was simplified by using the sysconfig module. The general
invocation of libnemo-python.so is now less error-prone.