git: Add nemo-git
This is a new extension providing an InfoBar which displays the current
git branch and a file changes summary whenever a user is inside a git
(sub-)directory. Also, there is a new column which shows the file status
according to git. So this allows seeing the git branch and untracked,
new, removed or renamed files at a glance.