gtkhash: Fix compilation error with libcrypto
When building with libcrypto enabled, we can encounter an error due to
OpenSSL is often built with MDC2 algorithm support disabled. To avoid
it, we now check for OpenSSL's OPENSSL_NO_MDC2 being set and if so, MDC2
is disabled by returning NULL instead of the actual function call.

Besides, nettle support is now enabled by default. This adds SHA3
hashing.

This fixes linuxmint/nemo-extensions#149