nemo-media-columns: Cache info and serve it synchronously when possible
In async mode (i.e. when using Nemo.OperationResult.IN_PROGRESS),
the view won't refresh the item if we don't call invalidate_extension_info().

Calling this results in Nemo getting stuck in what looks like an infinite recursion.
This is reproducible by browsing back and forth very quickly a volume which is slow
to load (and HDD for instance).

To prevent this from happening, this commit caches the info, so the first time
it is fetched, it is given back asynchronously, but after that is it given back
synchronously and without calling invalidate_extension_info().

The drawback is that the info might become obsolete if the file changes, or that
we might use too much memory by caching all that info.

If we want to fix that in the future, we can call del self.files_extension_info[uri]
after serving the info synchronously.

It might be sufficient and skip one call every two calls.. thus probably skipping
the callback resulting from calling invalidate_extension_info().