Add nemo-creation-date extension
Will build with or without actual kernel support for statx.

Will gracefully fail (and print a message) when running on a kernel
without support, and will just work when there is support (regardless
of whether it was compiled under a supporting kernel or not.)

Uses meson, as well as modern gobject macros to clean up the boilerplate
plugin code.