nemo-media-columns: Don't allow indefinite blocking when a bad file is encountered.
Using stopit allows limiting how long this extension gets to run
before it's forced to return.

Refactored a bit to make this thread-safe.

The overall impact of this extension may be reduced by changing
nemo as explained in code comments. We could avoid adding attributes
for *every* file, whether it's a relevant mimetype or not.

Ref: https://github.com/linuxmint/nemo-extensions/issues/348