simplify meson option handling
meson has the concept of "feature" options. Rather than either enabling
or disabling an optional dependency+feature via true/false boolean type,
a feature option accepts "enabled", "disabled", and "auto". By passing
the feature option directly to the dependency() lookup, meson knows
when/how to look it up without if/else soup.

The new functionality here is:
- auto type will try to detect if the dependency is available and use it
- disabled type will log "Dependency avahi-glib skipped: feature sharing disabled"

For equivalence with the previous option values, the default options
continue to force enable or force disable the features. No use of "auto"
(but people building the project can configure it that way, if they wish).